relocated all application http requests to own file
